+++ src/identifier.rs
@@ -36,8 +36,8 @@ const DECODE_MAP: [u8; 256] = [
 /// let x = "036Z968FU2TUGY7SVKFZNEWKK".parse::<Scru128Id>()?;
 /// assert_eq!(x.to_string(), "036Z968FU2TUGY7SVKFZNEWKK");
 ///
-/// let y = Scru128Id::from_u128(0x017fa1de51a80fd992f9e8cc2d5eb88e);
-/// assert_eq!(y.as_u128(), 0x017fa1de51a80fd992f9e8cc2d5eb88e);
+/// let y = Scru128Id::from(0x017fa1de51a80fd992f9e8cc2d5eb88eu128);
+/// assert_eq!(y.to_u128(), 0x017fa1de51a80fd992f9e8cc2d5eb88eu128);
 /// # Ok::<(), scru128::ParseError>(())
 /// ```
 #[derive(Copy, Clone, Eq, PartialEq, Ord, PartialOrd, Hash, Debug, Default)]
@@ -50,12 +50,17 @@ pub struct Scru128Id(u128);
 
 impl Scru128Id {
     /// Creates an object from a 128-bit unsigned integer.
+    ///
+    /// Use `Scru128Id::from(u128)` instead out of `const` context. This constructor may be
+    /// deprecated in the future once [const trait impls] are stabilized.
+    ///
+    /// [const trait impls]: https://github.com/rust-lang/rust/issues/67792
     pub const fn from_u128(int_value: u128) -> Self {
         Self(int_value)
     }
 
     /// Returns the 128-bit unsigned integer representation.
-    pub const fn as_u128(&self) -> u128 {
+    pub const fn to_u128(&self) -> u128 {
         self.0
     }
 
@@ -154,13 +159,13 @@ impl fmt::Display for Scru128Id {
 
 impl From<u128> for Scru128Id {
     fn from(value: u128) -> Self {
-        Self::from_u128(value)
+        Self(value)
     }
 }
 
 impl From<Scru128Id> for u128 {
     fn from(object: Scru128Id) -> Self {
-        object.as_u128()
+        object.to_u128()
     }
 }
 
@@ -228,11 +233,11 @@ mod tests {
 
             assert_eq!(from_fields, from_string);
             assert_eq!(
-                from_fields.as_u128(),
+                from_fields.to_u128(),
                 u128::from_str_radix(e.1, 36).unwrap()
             );
             assert_eq!(
-                from_string.as_u128(),
+                from_string.to_u128(),
                 u128::from_str_radix(e.1, 36).unwrap()
             );
             assert_eq!(
@@ -305,7 +310,9 @@ mod tests {
 
         for e in cases {
             assert_eq!(e.to_string().parse::<Scru128Id>(), Ok(e));
-            assert_eq!(Scru128Id::from_u128(e.as_u128()), e);
+            assert_eq!(Scru128Id::try_from(String::from(e)), Ok(e));
+            assert_eq!(Scru128Id::from_u128(e.to_u128()), e);
+            assert_eq!(Scru128Id::from(u128::from(e)), e);
             assert_eq!(
                 Scru128Id::from_fields(e.timestamp(), e.counter_hi(), e.counter_lo(), e.entropy()),
                 e